backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.7
langchain-core==0.3.19
langchain-groq==0.2.1
chromadb==0.5.20
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
requests==2.33.1
google-genai==1.14.0
